Boxplot of spikes within a window (optionally less a baseline)

Boxplot of spikes within a window (optionally less a baseline)

PlotOdourResponseFromSpikes(spiketimes, responseWindow, baselineWindow = NULL,
  freq = FALSE, PLOTFUN = stripchart, ...)

Arguments

spiketimes
list of spiketimes collected by CollectSpikesFromSweeps

responseWindow
vector of start and end time of odour response (in ms)

baselineWindow
vector of start and end time of baseline period (in ms)

freq
Plot spike rate in Hz rather than counts (default FALSE)

PLOTFUN
stripchart, boxplot or similar function (default stripchart)

...
Additional arguments passed on to PLOTFUN

Value

results of plotfun (if any)

See also

CollectSpikesFromSweeps

Other OdourResponse: OdourResponseFromSpikes

Examples

# NOT RUN {
spikes=CollectSpikesFromSweeps(
  system.file('igor','spikes','nm20110914c4',package='gphys'),
  subdir='BLOCKI',sweeps=0:4)
## stripchart
PlotOdourResponseFromSpikes(spikes,c(2200,2700),c(0,2000),pch=19,method='jitter',
 col=1:6)
## boxplot, in Hz
PlotOdourResponseFromSpikes(spikes,c(2200,2700),c(0,2000),freq=TRUE,
 PLOTFUN=boxplot)
# }