Take the mean of all the columns in a multi-time series (mts) object

Take the mean of all the columns in a multi-time series (mts) object

# S3 method for mts
mean(x, ...)

Arguments

x
an mts object

...
additional arguments (currently ignored)

Value

return value Single time series (ts)

See also

ts

Examples

require(IgorR)
w<-read.pxp(system.file("igor","WedJul407c2_001.pxp",package="IgorR"))
wts<-WaveToTimeSeries(w[c('RecordA0',"RecordA1")])
wm<-mean(wts)
summary(wm)

#>     Min.  1st Qu.   Median     Mean  3rd Qu.     Max. 
#> -207.219  -28.945  -19.703  -19.072   -9.273  162.000