Merge two spiketimes lists with different ODD configs

Merge two spiketimes lists with different ODD configs

# S3 method for spiketimes
merge(x, y, ...)

Arguments

x, y
spiketimes objects to merge

...
additional params (currently ignored)

Value

new spiketimes object with original attributes merged as appropriate and additional mergedwaveinfo attribute indicating origin of waves.

See also

+.spiketimes for combining repeats from the same ODD config

Other spiketimes: +.spiketimes, [.spiketimes, divide, is.spiketimes, spiketimes, split.spiketimes, subset.spiketimes